Change type of size variable in readBytes
size should have type unsigned long instead of int in order
to avoid overflows and lossy casts when passing U30 integers.

This commit fixes CVE-2017-9989 (fixes #86).

+++ util/read.h
@@ -20,7 +20,7 @@ long readSInt32(FILE *f);
 unsigned long readEncUInt32(FILE *f);
 unsigned long readEncUInt30(FILE *f);
 long readEncSInt32(FILE *f);
-char *readBytes(FILE *f,int size);
+char *readBytes(FILE *f,unsigned long size);
 char *readString(FILE *f);
 char *readSizedString(FILE *f,int size);
 double readDouble(FILE *f);
